Fix the issue of Peripheral processor synchronization queue error.
This error occurs because the queue is full and new data cannot be put into the queue.

The root cause is that GpsTime < requestTs < timestamp in GpsSource.getData. Because latest timestamp > requests, new data point will not be put into oldData from queue. And because GpsTime < requests, getData return None indicating that requestTs > the latest timestamp, which is not True. In syncSources, getData will be called again and again so the whole program falls into an endless loop.

Note: current solution is not perfect!!! When ts<requestTs<GpsTime, we need to do extrapolation (not interpolation). The extrapolation could be dangerous if GpsTime is far away from requestTs