Fix a bug in SampleManager RPC call.
From Derek:
The old one did not have the last line, but it needs to be in it in order for the ssim software to not mess with the pressure locking once it calls this function. A customer is currently seeing the same symptom that once the SSIM software runs and is finished, the pressure says unlocked even though the pressure is perfectly stable.