Add DebugMode option to supervisor.ini.
If Supervisor.py is run with python -O or -OO, the optimization
is propagated to the child processes.  This can be overriden
(and rpdb2 loaded) by putting 'DebugMode = True' in [GlobalDefaults]
of supervisor.ini.  Or you enable debugging of selected child
processes by putting 'DebugTrue = True' in the section where
the child process is defined.