Fix alarm setting panel exception when clicked.
When clicking on the alarm name or LED, it should bring
up a dialog to allow the settings to be changed. When this
is done, it calls Close() on a tooltip that no longer exist.
This throws an exception that prevents the dialog from
being created.  Now check to see if the tooltip exists and
is visible before calling Close().