Change temperature text labels in QuickGui so the labels would appear.
wxPython on Xubuntu can't render unicode, or at least I can't figure it out because
there isn't any good documentation on how to do it.  I sort of got it work on
main panel labels with wx.lib.StaticFancyText() but then it messes up the plot
labels.  The compromise is not to use the degree symbol and to instead spell out
"Celsius".