Working on alarm QuickGui display.
During warm up the system blinks yellow and transitions to green
when starting measurment mode.  The gas concentrations leds are
in the off state until measuring starts then go to red if the
cross the user set threshold.

BTW, this code is crap but that's because the foundation is dying under
enormous technical debt.