Changed builder to restore missing *.py files in Host/pydCaller.
Yuan removed the *.py files in Host/pydCaller to hide more of the
code.  These were replaced by *.pyo files.  The code works but
the ini files all have to be changed to run *.pyo instead of *.py.
I have changed the builder so that both *.py and *.pyo are in
pydCaller so existing config files still work.