Change default value of filter heater resistance
If the cable from filter heater board to logic board is not installed, the default value will be reported. Old value corresponds to 25 C which makes heater staying at maximum power. New value is closer to open circuit so it should mean a low temperature and sentry will be triggered.