Fix UserAdmin tool, didn't have anyway to exit the code.
The tool lets you log in but when you log out it goes back to the log in
screen and there, there is not button to exit the code.  Changed the
log in CANCEL button and the LOG OFF buttons to exit the code.  Also
made the main window frameless.