Fix DynamicNoiseFilter failure to load its ini file.
The search path was hard coded to look for the ini file in the local user's git
directory.  The code fails on a shippable analyzer where git doesn't exist. Hard
code it to look in /home/picarro/SI2000/AppConfig... yes this is a kludge but we
need this to work for RTM.