Fix wxPython error. See below.
If you see this error in the traceback

wx._core.PyAssertionError: C++ assertion "m_window" failed at ../src/gtk/dcclient.cpp(2043) in DoGetSize(): GetSize() doesn't work without window

You are trying to write to a window that hasn't been instantiated.
This bug is only seen if you are running X11 with GTK and you call
a wxWidget update in an __init__() before the main window has been
instantiated.  Here the fix is to eliminate the premature Update()
calls.