Fix key error exceptions in QuickGui.
In a couple of places QuickGui tries to read data from
GraphSequences that aren't created until after the system
has been running for awhile.  The code runs but prints
a bunch of exceptions in the QuickGui xterm. I put in
proper checks and exception handling to remove these harmless
error messages.