Fix I2-368, corrupt warmbox cal file.
This address the bug where an '8' was appended to the wb cal
file making it unusable.  Added sanity checks to the input
and output cal files. Removed threading in the update as a
race condition there probably was the root cause.  This fix
needs testing on an analyzer to confirm performance is the
same.