Delete ref. gas settings that don't help.
Ref. gas options in the validation ini are simplified in
the following way.
* Only one gas component can be defined.
* Eliminate the Desc and Vendor fields.

These eliminations were done to reduce the number of rows
in the ref. gas spreadsheet.  Only one gas component is shown
as we only measure one component anyway.  The description
and vendor fields are not necessary.  End users can put this
information in the Name field.