Added cell widget to restrict ref. gas accuracy.
Converted ref. gas accuracy from ppm to % as per Gerald's
recommendation.  % accuracy is how the vendor reports
their gases. The spreadsheet allows the user to set the
ref. gas accuracy to the most common values of 0.5, 1,
2, 5, 10 %.  Unspecified is "unk".  The setting passed
through to the final report but the setting is not yet used
in pass/fail test.