Important changes to the zero-air option.
Zero-air is now selectable with 3 options:
* No
* Standard
* Ultra

No means treat the gas as normal including the % deviation
test. Standard means it is zero-air with the gas in question
rated at < 1 ppm.  Ultra means the gas in question is
< 0.1 ppm.  These numbers are from the AirGas literature.
The zero-air test uses these numbers for the high threshold.
10 ppb is the bottom.

Reformatted the report data with the terminaltables module.