Partial fix I2-807, UNDO/SAVE confusion.
This fixes the buttons for the reference gas editor.  The UNDO
and SAVE buttons start out as disabled.  If any field is changed
the two buttons are enabled.  If either button is then clicked,
the undo or save action is done and the buttons then become
disabled indicating that the desired action has been accomplished.