Fix I2-878, I2-879, I2-880.
I2-878, still only accepts PPM but added the Data_Source_Units
option to the ini file to convert % conc or ppb data steams
to ppm so the tool still works.  All validataion reports
are in ppm.

I2-878, the report can still fail if the user selects a one
point validation and the user picks a ref. gas with the
accuracy set to UNK but now the report indicates an error
and tells the user to consult the manual.  In the gas editor,
when a user selects 1 point validation they are warned to
pick a gas with a defined accuracy.