ini and archiver refactoring.
Ported over Rev. D changes to the ini and file output paths
so that all output file now end up in /home/picarro/I2000/Log.
Previously some ended up in ~/git/host/src/main/python and
others in ~/git/host/Config/<type>/Log.

Changed Archiver so that dated directories are flattend.
2018/08/14 now become 2018-08-14/. Also the dated directories
are the primary directory and file types are subdirectories
of the dated directory so that it is easy to find all the files
of a given date.  RDFs are held in special dated directories
with RDF in the directory name (2018-08-14-RDF) so that it is
easy to delete the larger RDFs without sacrificing the dat files.