Increment version to 1.1.4, turn pressure reads on at 40Hz
Pressure read period is 25ms, which gives a 40Hz read rate.  MPR read
delay is 6ms, which can grow if we need to improve stability.  Kludge
delay is 3ms.  It seems like this could be small compared to the read
delay we need.  I'm not sure why this needs to be 3ms while the read
delay has to be 6ms...unless maybe the MPRs need more than 5ms.